+++ src/main/java/no/kristiania/database/Task.java
@@ -2,16 +2,16 @@ package no.kristiania.database;
 
 public class Task {
     private String name;
-    private Long id;
+    private int id;
     private String ColorCode;
 
     public String getName() { return name; }
 
     public void setName(String name) { this.name = name; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
 
     public String getColorCode() { return ColorCode; }
